# static/swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "This is the basic documentation for the Event manager API",
    "version": "1.0.0",
    "title": "Swagger Event manager",
    "contact": {
      "email": "kiocone@gmail.com"
    },
    "license": {
      "name": "Apache 2.0",
      "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
    }
  },
  "host": "localhost:5000",
  "basePath": "/v1",
  "tags": [
    {
      "name": "event",
      "description": "events information"
    }
  ],
  "schemes": [
    "http"
  ],
  "paths": {
    "/events": {
      "get": {
        "tags": [
          "event"
        ],
        "summary": "List Events",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "query",
            "name": "managed",
            "type": "boolean",
            "description": "filter by event managed"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "type": "array",
              "items": {
                "$ref": "#/definitions/Event"
              }
            }
          },
          "404": {
            "description": "No events found"
          }
        }
      },
      "post": {
        "tags": [
          "event"
        ],
        "summary": "Add a new event to the manager app",
        "description": "",
        "operationId": "addEvent",
        "consumes": [
          "application/json",
          "application/xml"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "Event object that needs to be added to the manager",
            "required": true,
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      }
    },
    "/events/{id}": {
      "get": {
        "tags": [
          "event"
        ],
        "summary": "Find event by ID",
        "description": "Returns a single event",
        "operationId": "getPetById",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of event to return",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "200": {
            "description": "successful operation",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          },
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Event not found"
          }
        }
      },
      "put": {
        "tags": [
          "event"
        ],
        "summary": "Updates a event in the manager with form data",
        "description": "",
        "operationId": "updateEvent",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of event that needs to be updated",
            "required": true,
            "type": "integer",
            "format": "int64"
          },
          {
            "in": "body",
            "name": "body",
            "description": "Updated of the event",
            "required": false,
            "type": "string",
            "schema": {
              "$ref": "#/definitions/Event"
            }
          }
        ],
        "responses": {
          "405": {
            "description": "Invalid input"
          }
        }
      },
      "delete": {
        "tags": [
          "event"
        ],
        "summary": "Deletes a event",
        "description": "",
        "operationId": "deletePet",
        "produces": [
          "application/json",
          "application/xml"
        ],
        "parameters": [
          {
            "name": "api_key",
            "in": "header",
            "required": false,
            "type": "string"
          },
          {
            "name": "id",
            "in": "path",
            "description": "Event id to delete",
            "required": true,
            "type": "integer",
            "format": "int64"
          }
        ],
        "responses": {
          "400": {
            "description": "Invalid ID supplied"
          },
          "404": {
            "description": "Event not found"
          }
        }
      }
    }
  },
  "definitions": {
    "ApiResponse": {
      "type": "object",
      "properties": {
        "code": {
          "type": "integer",
          "format": "int32"
        },
        "type": {
          "type": "string"
        },
        "message": {
          "type": "string"
        }
      }
    },
    "Category": {
      "type": "object",
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "name": {
          "type": "string"
        }
      },
      "xml": {
        "name": "Category"
      }
    },
    "Event": {
      "type": "object",
      "required": [
        "description",
        "type"
      ],
      "properties": {
        "id": {
          "type": "integer",
          "format": "int64"
        },
        "description": {
          "type": "string"
        },
        "type": {
          "type": "string",
          "example": "1 | 2 | 3"
        },
        "event_date": {
          "type": "string",
          "format": "date"
        },
        "viewed": {
          "type": "boolean",
          "description": "event reviewed in the manager"
        },
        "managed": {
          "type": "boolean",
          "description": "event reviewed in the manager"
        },
        "deleted": {
          "type": "boolean",
          "description": "event reviewed in the manager"
        }
      },
      "xml": {
        "name": "Event"
      }
    }
  },
  "externalDocs": {
    "description": "Find out more about Swagger",
    "url": "http://swagger.io"
  }
}